+++ src/include/numbers.h
@@ -13,4 +13,4 @@ size_t geti_digits(long n);
 /**
  * @brief returns the number of digits in a number
  */
-size_t getf_digits(float f, int decimals);
+size_t getf_digits(double f, int decimals);
